requirements.txt
discord.py[voice]==1.5.1
wikipedia==1.4.0
eyeD3==0.9.5
Pillow==8.0.0
redis==3.5.3
holidays==0.10.3
Flask==1.1.2
gunicorn==20.0.4
Authlib==0.14.1
python-dotenv==0.14.0
hiredis==1.1.0
sentry-sdk==0.19.1
pandas==1.1.3
blinker==1.4
pydub==0.24.1